# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: DEAN Orchestration API
  description: |
    The DEAN (Distributed Evolutionary Agent Network) API provides comprehensive endpoints for managing
    evolutionary optimization of code through autonomous agents. This API enables creation and management
    of agents, evolution trials, pattern discovery, and system monitoring.
  version: 1.0.0
  contact:
    name: DEAN Support Team
    email: support@dean-orchestration.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.dean-orchestration.com
    description: Production server
  - url: https://staging.dean-orchestration.com
    description: Staging server
  - url: http://localhost:8082
    description: Local development
security:
  - bearerAuth: []
...
paths:
...
  /api/evolution/start:
    post:
      tags:
        - Evolution
      summary: Start evolution trial
      description: Begin a new evolution trial with specified parameters
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StartEvolutionRequest'
            examples:
              standard:
                value:
                  repository: my-optimization-project
                  generations: 10
                  population_size: 50
                  mutation_rate: 0.15
                  crossover_rate: 0.7
                  fitness_threshold: 0.9
                  selection_method: tournament
      responses:
        '201':
          description: Evolution trial started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EvolutionTrial'
        '400':
          $ref: '#/components/responses/BadRequestError'
  /api/evolution/trials:
    get:
      tags:
        - Evolution
      summary: List evolution trials
      description: Get all evolution trials with optional filtering
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [running, completed, failed, cancelled]
          description: Filter by trial status
        - in: query
          name: start_date
          schema:
            type: string
            format: date-time
          description: Filter trials started after this date
        - in: query
          name: end_date
          schema:
            type: string
            format: date-time
          description: Filter trials started before this date
      responses:
        '200':
          description: List of evolution trials
          content:
            application/json:
              schema:
                type: object
                properties:
                  trials:
                    type: array
                    items:
                      $ref: '#/components/schemas/EvolutionTrial'
                  total:
                    type: integer
...
  /api/patterns:
    get:
      tags:
        - Patterns
      summary: List discovered patterns
      description: Retrieve patterns discovered during evolution trials
      parameters:
        - in: query
          name: min_confidence
          schema:
            type: number
            format: float
            minimum: 0
            maximum: 1
            default: 0.7
          description: Minimum confidence score
        - in: query
          name: category
          schema:
            type: string
            enum: [optimization, refactoring, security, performance]
          description: Pattern category
        - in: query
          name: language
          schema:
            type: string
            enum: [python, javascript, java]
          description: Programming language
      responses:
        '200':
          description: List of patterns
          content:
            application/json:
              schema:
                type: object
                properties:
                  patterns:
                    type: array
                    items:
                      $ref: '#/components/schemas/Pattern'
                  total:
                    type: integer
...
  /ws:
    get:
      tags:
        - WebSocket
      summary: WebSocket connection
      description: |
        Establish WebSocket connection for real-time updates.
        
        **WebSocket Events:**
        
        Sent by server:
        - `evolution.progress`: Evolution trial progress updates
        - `pattern.discovered`: New pattern discovered
        - `agent.status`: Agent status changes
        - `system.metrics`: Real-time system metrics
        
        Sent by client:
        - `subscribe`: Subscribe to specific event types
        - `unsubscribe`: Unsubscribe from event types
        - `ping`: Keep connection alive
      security:
        - bearerAuth: []
      responses:
        '101':
          description: Switching Protocols - WebSocket connection established
        '401':
          $ref: '#/components/responses/UnauthorizedError'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    StartEvolutionRequest:
      type: object
      required:
        - repository
        - generations
      properties:
        repository:
          type: string
          description: Repository or project to optimize
          example: my-optimization-project
        generations:
          type: integer
          minimum: 1
          maximum: 100
          description: Number of generations to evolve
          example: 10
        population_size:
          type: integer
          minimum: 10
          maximum: 1000
          default: 50
          description: Size of agent population
        mutation_rate:
          type: number
          format: float
          minimum: 0
          maximum: 1
          default: 0.15
          description: Probability of mutation
        crossover_rate:
          type: number
          format: float
          minimum: 0
          maximum: 1
          default: 0.7
          description: Probability of crossover
        fitness_threshold:
          type: number
          format: float
          minimum: 0
          maximum: 1
          default: 0.9
          description: Target fitness score
        selection_method:
          type: string
          enum: [tournament, roulette, rank]
          default: tournament
          description: Selection algorithm
    EvolutionTrial:
      type: object
      properties:
        id:
          type: string
          format: uuid
        repository:
          type: string
        status:
          type: string
          enum: [running, completed, failed, cancelled]
        current_generation:
          type: integer
        total_generations:
          type: integer
        best_fitness:
          type: number
          format: float
        average_fitness:
          type: number
          format: float
        patterns_discovered:
          type: integer
        started_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
...
    Pattern:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: Async Handler Optimization
        category:
          type: string
          enum: [optimization, refactoring, security, performance]
        confidence:
          type: number
          format: float
          minimum: 0
          maximum: 1
          example: 0.92
        impact_score:
          type: number
          format: float
          minimum: 0
          maximum: 1
          example: 0.85
        description:
          type: string
          example: Convert synchronous operations to async for improved performance
        language:
          type: string
          enum: [python, javascript, java]
        discovered_at:
          type: string
          format: date-time
        usage_count:
          type: integer
          description: Number of times pattern has been applied
...
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
        code:
          type: string
          description: Error code
        details:
          type: object
          description: Additional error details
  responses:
    UnauthorizedError:
      description: Authentication information is missing or invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Unauthorized
            code: AUTH_001
            details:
              message: Invalid or expired token
    BadRequestError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Bad Request
            code: VAL_001
            details:
              field: generations
              message: Value must be between 1 and 100
...