# service_stubs/evolution/main.py
#!/usr/bin/env python3
"""
Evolution API Stub

Provides a minimal implementation of the Evolution API for development.
Simulates evolution trials with realistic timing and state changes.
"""

import asyncio
import os
import random
import time
import json
from datetime import datetime, timedelta
from typing import Dict, List, Optional, Any, Set
from uuid import uuid4
from enum import Enum

from fastapi import FastAPI, HTTPException, WebSocket, WebSocketDisconnect, Depends, Header
from fastapi.middleware.cors import CORSMiddleware
from fastapi.security import HTTPBearer, HTTPAuthorizationCredentials
from pydantic import BaseModel, Field
import structlog
import uvicorn
import jwt

# Configure logging
logger = structlog.get_logger()

# Create FastAPI app
app = FastAPI(
    title="Evolution API Stub",
    description="Mock Evolution service for DEAN development",
    version="1.0.0"
)

# Add CORS middleware
app.add_middleware(
    CORSMiddleware,
    allow_origins=["*"],
    allow_credentials=True,
    allow_methods=["*"],
    allow_headers=["*"],
)

# Enums
class TrialStatus(str, Enum):
    PENDING = "pending"
    INITIALIZING = "initializing"
    RUNNING = "running"
    COMPLETED = "completed"
    FAILED = "failed"
    CANCELLED = "cancelled"

class PatternType(str, Enum):
    OPTIMIZATION = "optimization"
    REFACTORING = "refactoring"
    BUG_FIX = "bug_fix"
    FEATURE = "feature"
    ARCHITECTURE = "architecture"
    PERFORMANCE = "performance"

# In-memory storage
class EvolutionDataStore:
    def __init__(self):
        self.trials: Dict[str, Dict] = {}
        self.patterns: List[Dict] = []
        self.results: Dict[str, Dict] = {}
        self.active_websockets: Set[WebSocket] = set()
        self._generate_initial_patterns()
        
    def _generate_initial_patterns(self):
        """Generate some initial patterns."""
        pattern_templates = [
            ("Error handling pattern", PatternType.BUG_FIX, "Consistent error handling across services"),
            ("Async optimization", PatternType.OPTIMIZATION, "Convert sync operations to async for better performance"),
            ("Repository pattern", PatternType.ARCHITECTURE, "Implement repository pattern for data access"),
            ("Caching strategy", PatternType.PERFORMANCE, "Add caching layer for frequently accessed data"),
            ("API versioning", PatternType.FEATURE, "Implement API versioning strategy"),
            ("Logging standardization", PatternType.REFACTORING, "Standardize logging across all services")
        ]
        
        for i, (name, ptype, desc) in enumerate(pattern_templates):
            self.patterns.append({
                "id": f"pattern-base-{i}",
                "name": name,
                "type": ptype,
                "description": desc,
                "confidence": round(random.uniform(0.7, 0.95), 2),
                "occurrences": random.randint(10, 100),
                "impact_score": round(random.uniform(0.6, 0.9), 2),
                "discovered_at": (datetime.utcnow() - timedelta(days=random.randint(1, 30))).isoformat(),
                "repositories": [f"repo-{j}" for j in range(random.randint(1, 5))]
            })

# Initialize data store
data_store = EvolutionDataStore()

# Authentication configuration
JWT_SECRET = os.getenv("JWT_SECRET_KEY", "test-secret-key")  # Use same secret as main app
JWT_ALGORITHM = "HS256"

# Security scheme
security = HTTPBearer()

# Authentication dependency
def verify_token(credentials: HTTPAuthorizationCredentials = Depends(security)):
    """Verify JWT token for protected endpoints.
    
    This is a simplified version for the stub. In production, this would:
    - Validate token signature
    - Check token expiration
    - Verify user roles
    - Handle service-to-service authentication
    """
    token = credentials.credentials
    
    try:
        # Decode and verify token
        payload = jwt.decode(token, JWT_SECRET, algorithms=[JWT_ALGORITHM])
        
        # Check required fields
        if "sub" not in payload or "username" not in payload:
            raise HTTPException(status_code=401, detail="Invalid token structure")
            
        # Check token type
        if payload.get("type") not in ["access", "service"]:
            raise HTTPException(status_code=401, detail="Invalid token type")
            
        return payload
        
    except jwt.ExpiredSignatureError:
        raise HTTPException(status_code=401, detail="Token has expired")
    except jwt.InvalidTokenError:
        raise HTTPException(status_code=401, detail="Invalid token")
    except Exception as e:
        logger.error("Token verification failed", error=str(e))
        raise HTTPException(status_code=401, detail="Authentication failed")

def get_current_user(token_data: dict = Depends(verify_token)):
    """Get current user from token."""
    return {
        "user_id": token_data.get("sub"),
        "username": token_data.get("username"),
        "roles": token_data.get("roles", [])
    }

# Pydantic models
class EvolutionConfig(BaseModel):
    repository: str
    generations: int = Field(default=10, ge=1, le=100)
    population_size: int = Field(default=20, ge=5, le=100)
    mutation_rate: float = Field(default=0.1, ge=0.0, le=1.0)
    crossover_rate: float = Field(default=0.7, ge=0.0, le=1.0)
    fitness_threshold: float = Field(default=0.9, ge=0.0, le=1.0)
    max_runtime_minutes: int = Field(default=30, ge=1, le=120)
    parameters: Dict[str, Any] = Field(default_factory=dict)

class EvolutionTrial(BaseModel):
    trial_id: str
    repository: str
    config: EvolutionConfig
    status: TrialStatus
    started_at: str
    completed_at: Optional[str] = None
    current_generation: int = 0
    best_fitness: float = 0.0
    improvements: List[float] = Field(default_factory=list)
    patterns_discovered: int = 0
    error: Optional[str] = None

class EvolutionResult(BaseModel):
    trial_id: str
    repository: str
    status: TrialStatus
    generations_completed: int
    final_fitness: float
    total_improvements: float
    patterns_discovered: List[Dict]
    execution_time_seconds: float
    metadata: Dict[str, Any] = Field(default_factory=dict)

# WebSocket connection manager
class ConnectionManager:
    def __init__(self):
        self.active_connections: Set[WebSocket] = set()
        
    async def connect(self, websocket: WebSocket):
        await websocket.accept()
        self.active_connections.add(websocket)
        logger.info("WebSocket connected", total_connections=len(self.active_connections))
        
    def disconnect(self, websocket: WebSocket):
        self.active_connections.discard(websocket)
        logger.info("WebSocket disconnected", total_connections=len(self.active_connections))
        
    async def broadcast(self, message: dict):
        """Broadcast message to all connected clients."""
        if self.active_connections:
            message_str = json.dumps(message)
            disconnected = set()
            
            for connection in self.active_connections:
                try:
                    await connection.send_text(message_str)
                except:
                    disconnected.add(connection)
                    
            # Clean up disconnected clients
            self.active_connections -= disconnected

manager = ConnectionManager()

# Health endpoint
@app.get("/health")
async def health_check():
    """Health check endpoint."""
    return {
        "status": "healthy",
        "service": "evolution-stub",
        "timestamp": datetime.utcnow().isoformat(),
        "uptime_seconds": int(time.time() - app.state.start_time),
        "active_trials": len([t for t in data_store.trials.values() 
                            if t["status"] == TrialStatus.RUNNING])
    }

# Evolution endpoints
@app.post("/evolution/start")
async def start_evolution(config: EvolutionConfig, user: dict = Depends(get_current_user)):
    """Start a new evolution trial."""
    trial_id = f"trial-{uuid4().hex[:8]}"
    
    trial = EvolutionTrial(
        trial_id=trial_id,
        repository=config.repository,
        config=config,
        status=TrialStatus.INITIALIZING,
        started_at=datetime.utcnow().isoformat()
    )
    
    data_store.trials[trial_id] = trial.dict()
    
    # Start async evolution simulation
    asyncio.create_task(simulate_evolution(trial_id))
    
    logger.info("Evolution trial started", 
                trial_id=trial_id, 
                repository=config.repository,
                generations=config.generations,
                user=user["username"])
    
    return trial

async def simulate_evolution(trial_id: str):
    """Simulate evolution trial execution."""
    trial = data_store.trials[trial_id]
    
    try:
        # Initial delay
        await asyncio.sleep(random.uniform(1, 2))
        
        # Update to running
        trial["status"] = TrialStatus.RUNNING
        await broadcast_trial_update(trial_id, "Trial started")
        
        # Simulate generations
        config = trial["config"]
        initial_fitness = random.uniform(0.3, 0.5)
        current_fitness = initial_fitness
        
        for generation in range(config["generations"]):
            # Check if cancelled
            if trial["status"] == TrialStatus.CANCELLED:
                break
                
            # Simulate generation processing
            await asyncio.sleep(random.uniform(0.5, 2.0))
            
            # Update fitness (tends to improve)
            improvement = random.uniform(0.01, 0.1) * (1 - current_fitness)
            current_fitness = min(0.99, current_fitness + improvement)
            
            trial["current_generation"] = generation + 1
            trial["best_fitness"] = round(current_fitness, 3)
            trial["improvements"].append(round(improvement, 3))
            
            # Occasionally discover patterns
            if random.random() > 0.7:
                pattern = create_discovered_pattern(trial_id, generation)
                data_store.patterns.append(pattern)
                trial["patterns_discovered"] += 1
                
                await broadcast_trial_update(
                    trial_id, 
                    f"Pattern discovered in generation {generation + 1}",
                    {"pattern": pattern}
                )
            
            # Broadcast progress
            await broadcast_trial_update(
                trial_id,
                f"Generation {generation + 1} completed",
                {
                    "generation": generation + 1,
                    "fitness": current_fitness,
                    "improvement": improvement
                }
            )
            
        # Complete the trial
        trial["status"] = TrialStatus.COMPLETED
        trial["completed_at"] = datetime.utcnow().isoformat()
        
        # Create result
        result = EvolutionResult(
            trial_id=trial_id,
            repository=trial["repository"],
            status=TrialStatus.COMPLETED,
            generations_completed=trial["current_generation"],
            final_fitness=trial["best_fitness"],
            total_improvements=sum(trial["improvements"]),
            patterns_discovered=[p for p in data_store.patterns 
                               if p.get("trial_id") == trial_id],
            execution_time_seconds=(
                datetime.fromisoformat(trial["completed_at"]) - 
                datetime.fromisoformat(trial["started_at"])
            ).total_seconds(),
            metadata={
                "initial_fitness": initial_fitness,
                "improvement_rate": sum(trial["improvements"]) / len(trial["improvements"]) 
                                   if trial["improvements"] else 0
            }
        )
        
        data_store.results[trial_id] = result.dict()
        
        await broadcast_trial_update(trial_id, "Trial completed successfully")
        logger.info("Evolution trial completed", trial_id=trial_id)
        
    except Exception as e:
        trial["status"] = TrialStatus.FAILED
        trial["error"] = str(e)
        trial["completed_at"] = datetime.utcnow().isoformat()
        
        await broadcast_trial_update(trial_id, f"Trial failed: {str(e)}")
        logger.error("Evolution trial failed", trial_id=trial_id, error=str(e))

def create_discovered_pattern(trial_id: str, generation: int) -> Dict:
    """Create a newly discovered pattern."""
    pattern_types = list(PatternType)
    pattern_names = [
        "Async handler optimization",
        "Resource cleanup pattern",
        "Circuit breaker implementation",
        "Batch processing optimization",
        "Memory leak prevention",
        "Concurrent request handling"
    ]
    
    return {
        "id": f"pattern-{uuid4().hex[:8]}",
        "trial_id": trial_id,
        "name": random.choice(pattern_names),
        "type": random.choice(pattern_types),
        "description": f"Pattern discovered in generation {generation + 1}",
        "confidence": round(random.uniform(0.75, 0.95), 2),
        "occurrences": random.randint(1, 20),
        "impact_score": round(random.uniform(0.7, 0.95), 2),
        "discovered_at": datetime.utcnow().isoformat(),
        "generation": generation + 1,
        "repositories": [data_store.trials[trial_id]["repository"]]
    }

async def broadcast_trial_update(trial_id: str, message: str, data: Optional[Dict] = None):
    """Broadcast trial update to WebSocket clients."""
    update = {
        "type": "trial_update",
        "trial_id": trial_id,
        "message": message,
        "timestamp": datetime.utcnow().isoformat(),
        "data": data or {}
    }
    await manager.broadcast(update)

@app.get("/evolution/{trial_id}/status")
async def get_trial_status(trial_id: str, user: dict = Depends(get_current_user)):
    """Get the status of an evolution trial."""
    if trial_id not in data_store.trials:
        raise HTTPException(status_code=404, detail="Trial not found")
    
    return data_store.trials[trial_id]

@app.post("/evolution/{trial_id}/cancel")
async def cancel_trial(trial_id: str, user: dict = Depends(get_current_user)):
    """Cancel a running evolution trial."""
    if trial_id not in data_store.trials:
        raise HTTPException(status_code=404, detail="Trial not found")
    
    trial = data_store.trials[trial_id]
    
    if trial["status"] not in [TrialStatus.PENDING, TrialStatus.INITIALIZING, TrialStatus.RUNNING]:
        raise HTTPException(status_code=400, detail="Trial cannot be cancelled")
    
    trial["status"] = TrialStatus.CANCELLED
    trial["completed_at"] = datetime.utcnow().isoformat()
    
    await broadcast_trial_update(trial_id, "Trial cancelled by user")
    logger.info("Evolution trial cancelled", trial_id=trial_id, user=user["username"])
    
    return {"status": "cancelled", "trial_id": trial_id}

@app.get("/evolution/trials")
async def list_trials(
    repository: Optional[str] = None,
    status: Optional[TrialStatus] = None,
    limit: int = 100,
    offset: int = 0,
    user: dict = Depends(get_current_user)
):
    """List evolution trials."""
    trials = list(data_store.trials.values())
    
    # Filter by repository
    if repository:
        trials = [t for t in trials if t["repository"] == repository]
    
    # Filter by status
    if status:
        trials = [t for t in trials if t["status"] == status]
    
    # Sort by start time descending
    trials.sort(key=lambda t: t["started_at"], reverse=True)
    
    # Paginate
    total = len(trials)
    trials = trials[offset:offset + limit]
    
    return {
        "trials": trials,
        "total": total,
        "limit": limit,
        "offset": offset
    }

@app.get("/evolution/results")
async def get_results(
    repository: Optional[str] = None,
    min_fitness: Optional[float] = None,
    limit: int = 100,
    user: dict = Depends(get_current_user)
):
    """Get evolution results."""
    results = list(data_store.results.values())
    
    # Filter by repository
    if repository:
        results = [r for r in results if r["repository"] == repository]
    
    # Filter by minimum fitness
    if min_fitness is not None:
        results = [r for r in results if r["final_fitness"] >= min_fitness]
    
    # Sort by fitness descending
    results.sort(key=lambda r: r["final_fitness"], reverse=True)
    
    return {
        "results": results[:limit],
        "total": len(results)
    }

@app.get("/patterns")
async def get_patterns(
    pattern_type: Optional[PatternType] = None,
    min_confidence: float = 0.0,
    limit: int = 100,
    user: dict = Depends(get_current_user)
):
    """Get discovered patterns."""
    patterns = data_store.patterns
    
    # Filter by type
    if pattern_type:
        patterns = [p for p in patterns if p["type"] == pattern_type]
    
    # Filter by confidence
    patterns = [p for p in patterns if p["confidence"] >= min_confidence]
    
    # Sort by impact score descending
    patterns.sort(key=lambda p: p["impact_score"], reverse=True)
    
    return {
        "patterns": patterns[:limit],
        "total": len(patterns),
        "types": {ptype: len([p for p in data_store.patterns if p["type"] == ptype]) 
                  for ptype in PatternType}
    }

@app.get("/evolution/metrics")
async def get_evolution_metrics(user: dict = Depends(get_current_user)):
    """Get evolution system metrics."""
    trials = data_store.trials.values()
    completed_trials = [t for t in trials if t["status"] == TrialStatus.COMPLETED]
    
    avg_fitness = 0.0
    avg_generations = 0
    avg_patterns = 0.0
    
    if completed_trials:
        avg_fitness = sum(t["best_fitness"] for t in completed_trials) / len(completed_trials)
        avg_generations = sum(t["current_generation"] for t in completed_trials) / len(completed_trials)
        avg_patterns = sum(t["patterns_discovered"] for t in completed_trials) / len(completed_trials)
    
    return {
        "total_trials": len(trials),
        "active_trials": len([t for t in trials if t["status"] == TrialStatus.RUNNING]),
        "completed_trials": len(completed_trials),
        "failed_trials": len([t for t in trials if t["status"] == TrialStatus.FAILED]),
        "total_patterns": len(data_store.patterns),
        "average_fitness": round(avg_fitness, 3),
        "average_generations": round(avg_generations, 1),
        "average_patterns_per_trial": round(avg_patterns, 1)
    }

# WebSocket endpoint
@app.websocket("/ws")
async def websocket_endpoint(websocket: WebSocket):
    """WebSocket endpoint for real-time updates.
    
    Note: WebSocket authentication typically happens via:
    - Query parameter: ws://localhost:8083/ws?token=xxx
    - First message after connection
    - Subprotocol headers
    
    For this stub, we'll accept connections but log authentication status.
    """
    """WebSocket endpoint for real-time updates."""
    # Check for token in query parameters
    token = websocket.query_params.get("token")
    authenticated = False
    user_info = None
    
    if token:
        try:
            payload = jwt.decode(token, JWT_SECRET, algorithms=[JWT_ALGORITHM])
            authenticated = True
            user_info = {
                "user_id": payload.get("sub"),
                "username": payload.get("username")
            }
        except:
            logger.warning("WebSocket connection with invalid token")
    
    await manager.connect(websocket)
    
    try:
        # Send initial connection message
        await websocket.send_json({
            "type": "connection",
            "message": "Connected to Evolution API WebSocket",
            "authenticated": authenticated,
            "user": user_info["username"] if user_info else None,
            "timestamp": datetime.utcnow().isoformat()
        })
        
        # Keep connection alive
        while True:
            # Wait for any message from client (ping/pong)
            data = await websocket.receive_text()
            
            # Echo back as pong
            if data == "ping":
                await websocket.send_text("pong")
                
    except WebSocketDisconnect:
        manager.disconnect(websocket)
    except Exception as e:
        logger.error("WebSocket error", error=str(e))
        manager.disconnect(websocket)

# Startup event
@app.on_event("startup")
async def startup_event():
    """Initialize the service."""
    app.state.start_time = time.time()
    logger.info("Evolution API stub started", port=8083)
    
    # Create some sample completed trials
    for i in range(3):
        trial_id = f"trial-sample-{i}"
        config = EvolutionConfig(
            repository=f"sample-repo-{i}",
            generations=5,
            population_size=10
        )
        
        trial = EvolutionTrial(
            trial_id=trial_id,
            repository=config.repository,
            config=config,
            status=TrialStatus.COMPLETED,
            started_at=(datetime.utcnow() - timedelta(hours=i+1)).isoformat(),
            completed_at=(datetime.utcnow() - timedelta(hours=i)).isoformat(),
            current_generation=5,
            best_fitness=round(0.8 + i * 0.05, 2),
            improvements=[0.1, 0.08, 0.06, 0.04, 0.02],
            patterns_discovered=random.randint(1, 3)
        )
        
        data_store.trials[trial_id] = trial.dict()

if __name__ == "__main__":
    port = int(os.getenv("SERVICE_PORT", "8083"))
    uvicorn.run(
        "main:app",
        host="0.0.0.0",
        port=port,
        reload=True,
        log_config={
            "version": 1,
            "disable_existing_loggers": False,
            "formatters": {
                "default": {
                    "format": "%(asctime)s - %(name)s - %(levelname)s - %(message)s",
                },
            },
            "handlers": {
                "default": {
                    "formatter": "default",
                    "class": "logging.StreamHandler",
                    "stream": "ext://sys.stdout",
                },
            },
            "root": {
                "level": os.getenv("LOG_LEVEL", "INFO"),
                "handlers": ["default"],
            },
        }
    )